requirements.txt
streamlit
pandas
numpy
plotly
reportlab
openpyxl
matplotlib
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
statsmodels
scikit-learn
xlsxwriter